+++ api/app/clients/agents/methods/index.js
@@ -1,0 +1,9 @@
+const addImages = require('./addImages');
+const createLLM = require('./createLLM');
+const handleOutputs = require('./handleOutputs');
+
+module.exports = {
+  addImages,
+  createLLM,
+  ...handleOutputs,
+};
